+++ src/Api/Segments/Endpoints/ReportSegment.php
@@ -4,7 +4,9 @@ declare(strict_types=1);
 
 namespace Evgeek\Moysklad\Api\Segments\Endpoints;
 
+use Evgeek\Moysklad\Dictionaries\Endpoint;
+
 class ReportSegment extends AbstractEndpointSegmentNamed
 {
-    protected const SEGMENT = 'report';
+    protected const SEGMENT = Endpoint::REPORT;
 }