+++ src/Dictionaries/Entity.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Evgeek\Moysklad\Dictionaries;
+
+class Entity
+{
+    public const PRODUCT = 'product';
+    public const EMPLOYEE = 'employee';
+}