+++ src/Meta/MetaMaker.php
@@ -13,6 +13,16 @@ class MetaMaker
     {
     }
 
+    public function product(string $guid)
+    {
+        return Meta::product($guid, $this->formatter);
+    }
+
+    public function employee(string $guid)
+    {
+        return Meta::employee($guid, $this->formatter);
+    }
+
     public function create(array $path, string $type)
     {
         return Meta::create($path, $type, $this->formatter);